requirements.txt
biopython==1.73
cycler==0.10.0
Cython==0.29.6
kiwisolver==1.0.1
matplotlib==3.0.3
mock==2.0.0
numexpr==2.6.9
numpy==1.16.2
pbr==5.1.3
pyparsing==2.3.1
python-dateutil==2.8.0
scipy==1.2.1
six==1.12.0
tables==3.5.1
wheel==0.24.0
